refactor: extract form-layout logic into reusable mixins
