+++ packages/form-layout/src/vaadin-form-layout.d.ts
@@ -4,16 +4,10 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
-import { ResizeMixin } from '@vaadin/component-base/src/resize-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { FormLayoutMixin } from './vaadin-form-layout-mixin.js';
 
-export type FormLayoutLabelsPosition = 'aside' | 'top';
-
-export type FormLayoutResponsiveStep = {
-  minWidth?: string | 0;
-  columns: number;
-  labelsPosition?: FormLayoutLabelsPosition;
-};
+export * from './vaadin-form-layout-mixin.js';
 
 /**
  * `<vaadin-form-layout>` is a Web Component providing configurable responsive
@@ -105,51 +99,7 @@ export type FormLayoutResponsiveStep = {
  * ---|---|---
  * `--vaadin-form-layout-column-spacing` | Length of the spacing between columns | `2em`
  */
-declare class FormLayout extends ResizeMixin(ElementMixin(ThemableMixin(HTMLElement))) {
-  /**
-   * Allows specifying a responsive behavior with the number of columns
-   * and the label position depending on the layout width.
-   *
-   * Format: array of objects, each object defines one responsive step
-   * with `minWidth` CSS length, `columns` number, and optional
-   * `labelsPosition` string of `"aside"` or `"top"`. At least one item is required.
-   *
-   * #### Examples
-   *
-   * ```javascript
-   * formLayout.responsiveSteps = [{columns: 1}];
-   * // The layout is always a single column, labels aside.
-   * ```
-   *
-   * ```javascript
-   * formLayout.responsiveSteps = [
-   *   {minWidth: 0, columns: 1},
-   *   {minWidth: '40em', columns: 2}
-   * ];
-   * // Sets two responsive steps:
-   * // 1. When the layout width is < 40em, one column, labels aside.
-   * // 2. Width >= 40em, two columns, labels aside.
-   * ```
-   *
-   * ```javascript
-   * formLayout.responsiveSteps = [
-   *   {minWidth: 0, columns: 1, labelsPosition: 'top'},
-   *   {minWidth: '20em', columns: 1},
-   *   {minWidth: '40em', columns: 2}
-   * ];
-   * // Default value. Three responsive steps:
-   * // 1. Width < 20em, one column, labels on top.
-   * // 2. 20em <= width < 40em, one column, labels aside.
-   * // 3. Width >= 40em, two columns, labels aside.
-   * ```
-   */
-  responsiveSteps: FormLayoutResponsiveStep[];
-
-  /**
-   * Update the layout.
-   */
-  protected _updateLayout(): void;
-}
+declare class FormLayout extends FormLayoutMixin(ElementMixin(ThemableMixin(HTMLElement))) {}
 
 declare global {
   interface HTMLElementTagNameMap {
